refactor: move the definition of Renderer
